Changing classes names
