+++ src/main/java/edu/cmu/inmind/multiuser/socialreasoner/control/bn/BehaviorNetwork.java
@@ -1,17 +1,22 @@
 package edu.cmu.inmind.multiuser.socialreasoner.control.bn;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.Vector;
+import edu.cmu.inmind.multiuser.socialreasoner.control.SocialReasonerController;
+
+import java.text.DecimalFormat;
+import java.util.*;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.regex.Pattern;
 
 /**
  * @author oromero
  */
-public class BehaviorNetwork implements BehaviorNetworkInterface{
-	private List<BehaviorInterface> modules = new Vector<>();
-	private List <String> states = new Vector <>();
-	private List <String> goals = new Vector <>();
-	private List <String> goalsResolved = new Vector <>();
+public class BehaviorNetworkPlus{
+	private List<BehaviorPlus> modules = new Vector<>();
+//	private List<String> states = new Vector <>();
+    private CopyOnWriteArrayList<String> states = new CopyOnWriteArrayList<>();
+	private List<String> goals = new Vector <>();
+    private boolean removePrecond = false;
+	private transient List<String> goalsResolved = new Vector <>();
 
 	private double pi = 15; //20 the mean level of activation,
 	private double theta = 15; //45 the threshold of activation, where is lowered 10% every time no module could be selected, and is reset to its initial value whenever a module becomes active.
@@ -19,43 +24,71 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	private double phi = 70; //20 // 90 the amount of activation energy injected by the state per true proposition,
 	private double gamma = 20; //70 the amount of activation energy injected by the goals per goal,
 	private double delta = 50; // 90 not defined the amount of activation energy taken away by the protected goals per protected goal.
-	private double[][] activationSuccesors;
-	private double[][] activationPredeccesors;
-	private double[][] activationConflicters;
-	private double[] activationInputs;
-	private double[] activationLinks;
-	private boolean execution = false;
-	private int indexBehActivated = -1;
-	private int cont = 1;
-	Double[] activations;
 
-	public List<BehaviorInterface> getModules() {
+	private transient double[][] activationSuccesors;
+	private transient double[][] activationPredeccesors;
+	private transient double[][] activationConflicters;
+	private transient double[] activationInputs;
+	private transient double[] activationLinks;
+	private transient boolean execution = false;
+	private transient int indexBehActivated = -1;
+	private transient int cont = 1;
+	private transient Double[] activations;
+    private transient boolean verbose = false;
+    private transient List<String> previousStates;
+    private transient List<BehaviorPlus> modulesCopy = new Vector<>();
+    private transient String nameBehaviorActivated;
+	private String output;
+	private String matchesOutput;
+	private String statesOutput;
+
+	public BehaviorNetworkPlus(){}
+
+	public List<BehaviorPlus> getModules() {
 		return modules;
 	}
 
-	public Double[] getActivations() {
-		return activations;
+	public synchronized double[] getActivations() {
+		int size = activations.length;
+		double[] actvs = new double[ size ];
+		for( int i = 0; i < size; i++){
+			actvs[i] = activations[i];
+		}
+		return actvs;
 	}
 
 	public int getIdxBehActivated() {
 		return indexBehActivated;
 	}
 
-	public BehaviorInterface getBehaviorActivated(){
+    public boolean isRemovePrecond() {
+        return removePrecond;
+    }
+
+    public BehaviorPlus getBehaviorActivated(){
 		if( indexBehActivated >= 0 ) {
 			return modules.get(indexBehActivated);
 		}
 		return null;
 	}
 
-	public String getNameBehaviorActivated() {
-		if( indexBehActivated != -1 ){
-			return modules.get( indexBehActivated ).getName();
-		}
-		return "";
+	public String getOutput() {
+		return output;
 	}
 
-	public void setModules(List<BehaviorInterface> modules) {
+	public String getMatchesOutput() {
+		return matchesOutput;
+	}
+
+	public String getStatesOutput() {
+		return statesOutput;
+	}
+
+	public String getNameBehaviorActivated() {
+        return nameBehaviorActivated;
+	}
+
+	public void setModules(List<BehaviorPlus> modules) {
 		this.modules = modules;
 		activationSuccesors = new double[modules.size()][modules.size()];
 		activationPredeccesors = new double[modules.size()][modules.size()];
@@ -64,10 +97,10 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 		activationLinks = new double[modules.size()];
 	}
 
-	public void setModules(List<BehaviorInterface> modules, int size) {
+	public void setModules(List<BehaviorPlus> modules, int size) {
 		this.modules = modules;
 		for( int i = 0; i < modules.size(); i++ ){
-			modules.get(i).setId( i );
+			modules.get(i).setIdx(i);
 		}
 
 		activationSuccesors = new double[modules.size()][modules.size()];
@@ -135,32 +168,65 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * (the state of the environment as perceived by the agent) S being implemented
 	 * by an independent process (or the real world),
 	 */
-	public Collection<String> getState (){
+	public CopyOnWriteArrayList<String> getState (){
 		if(states != null)
 			return states;
-		return new Vector<>();
+		return new CopyOnWriteArrayList<>();
 	}
 
-	public void setState(List<String> states){
-		this.states = new Vector<>(states);
-	}
-
-	public void addState(List<String> states){
-		if(states == null)
-			this.states = null;
-		else
-			for(int i = 0; i < states.size(); i++) {
-				if( !this.states.contains(states.get(i)) ) {
-					this.states.add(states.get(i));
-				}
+	public String getStateString(){
+		String result = "";
+		if(states != null) {
+			for( String state : states ){
+				result += state + ", ";
 			}
+			result = result.substring(0, result.length()-2);
+		}
+		return result;
 	}
+
+    public String getPreviousStateString(){
+        String result = "";
+        if(previousStates != null) {
+            for( String state : previousStates ){
+                result += state + ", ";
+            }
+            result = result.substring(0, result.length()-2);
+        }
+        return result;
+    }
+
+	public synchronized void setState(CopyOnWriteArrayList<String> states){
+        try {
+            this.states = states; //new CopyOnWriteArrayList<>(states);
+        }catch (Exception e){
+            e.printStackTrace();
+        }
+	}
+
+	// Many components hold a reference to the states list. Update in place to ensure that all components are up to date
+	// TODO: Refactor so there is a single source of truth for BN state
+	public synchronized void updateState(List<String> states) {
+		getState().clear();
+		getState().addAll(states);
+	}
+
+//	public void addState(List<String> states){
+//		if(states == null)
+//			this.states = null;
+//		else
+//			for(int i = 0; i < states.size(); i++) {
+//				if( !this.states.contains(states.get(i)) ) {
+//					this.states.add(states.get(i));
+//				}
+//			}
+//	}
 
 	/**
 	 * a function G(t) returning the propositions that are a goal of the agent at time
 	 * t G being implemented by an independent process,
 	 */
-	public Collection<String> getGoals (){
+	public List<String> getGoals (){
 		if(goals != null)
 			return goals;
 		return new Vector<>();
@@ -199,16 +265,30 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * of S (t)), and 0 otherwise.
 	 */
 	public boolean executable (int i){
-		return modules.get(i).isExecutable(states);
+		try {
+			return modules.get(i).isExecutable(states);
+		}catch (Exception e){
+			e.printStackTrace();
+		}
+		return false;
 	}
+
+    public boolean executable (int idx, int maximum){
+        try {
+            return modules.get(idx).isExecutable(maximum);
+        }catch (Exception e){
+            e.printStackTrace();
+        }
+        return false;
+    }
 
 	/**
 	 * a function M (j), which returns the set of modules that match proposition j ,
 	 * i.e., the modules x for which j E cx,
 	 */
-	public List<BehaviorInterface> matchProposition (String proposition){
-		List<BehaviorInterface> behaviors = new Vector<>();
-		for(BehaviorInterface behavior : modules){
+	public List<BehaviorPlus> matchProposition(String proposition){
+		Vector<BehaviorPlus> behaviors = new Vector<>();
+		for(BehaviorPlus behavior : modules){
 			if(behavior.isPrecondition(proposition))
 				behaviors.add(behavior);
 		}
@@ -219,11 +299,11 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * a function A(j ), which returns the set of modules that achieve proposition j ,
 	 * i.e., the modules x for which j E ax,
 	 */
-	public List<BehaviorInterface> achieveProposition (String proposition){
-		List<BehaviorInterface> behaviors = new Vector<>();
-		for( BehaviorInterface beh : modules){
-			if(beh.isSuccesor(proposition))
-				behaviors.add(beh);
+	public Vector<BehaviorPlus> achieveProposition(String proposition){
+		Vector<BehaviorPlus> behaviors = new Vector<>();
+		for(BehaviorPlus beh : modules){
+			if( beh.isSuccesor(proposition) )
+				behaviors.add( beh );
 		}
 		return behaviors;
 	}
@@ -232,11 +312,11 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * a function U (j ), which returns the set of modules that undo proposition j , i.e.,
 	 * the modules x for which j E dx,
 	 */
-	public List<BehaviorInterface> undoProposition (String proposition){
-		List<BehaviorInterface> behaviors = new Vector<>();
-		for(BehaviorInterface beh : modules){
-			if(beh.isInhibition(proposition))
-				behaviors.add(beh);
+	public Vector<BehaviorPlus> undoProposition(String proposition){
+		Vector<BehaviorPlus> behaviors = new Vector<>();
+		for(BehaviorPlus beh : modules){
+			if( beh.isInhibition(proposition) )
+				behaviors.add( beh );
 		}
 		return behaviors;
 	}
@@ -245,13 +325,12 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * a function U (j ), which returns the set of modules that undo proposition j , i.e.,
 	 * the modules x for which j E dx, and j E S(t)
 	 */
-	public List<BehaviorInterface> undoPropositionState (String proposition, int indexBehx){
-		List<BehaviorInterface> behaviors = new Vector<>();
-		for(BehaviorInterface beh : modules){
-			if(beh.getId() != indexBehx){
-				BehaviorInterface behy = beh;
-				if(behy.isInhibition(proposition) && states.contains(proposition))
-					behaviors.add(behy);
+	public Vector<BehaviorPlus> undoPropositionState(String proposition, int indexBehx){
+		Vector<BehaviorPlus> behaviors = new Vector<>();
+		for(BehaviorPlus behx : modules){
+			if(behx.getIdx() != indexBehx){
+				if( behx.isInhibition(proposition) && states.contains(proposition) )
+					behaviors.add(behx);
 			}
 		}
 		return behaviors;
@@ -262,10 +341,10 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 / module is computed.
 	 */
 	public void computeActivation (){
-		List<String> statesList = new Vector<>(getState());
-		List<String> goalsList = new Vector<>(getGoals());
-		List<String> goalsRList = new Vector<>(getGoalsR());
-		int[] matchesByState = new int[states.size()];
+		Vector<String> statesList = new Vector<>(getState());
+		Vector<String> goalsList = new Vector<>(getGoals());
+		Vector<String> goalsRList = new Vector<>(getGoalsR());
+		int[] matchesByState = new int[statesList.size()];
 		int[] matchesByGoal = new int[goals.size()];
 		int[] matchesByGoalsResolved = new int[states.size()];//new int[goals.size()];
 
@@ -289,7 +368,6 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 			double actGoal = modules.get(i).calculateInputFromGoals(goalsList, matchesByGoal, gamma);
 			double actGoalR = modules.get(i).calculateTakeAwayByProtectedGoals(goalsRList, matchesByGoalsResolved, delta); //ojrl states
 			activationInputs[i] += actState + actGoal - actGoalR;
-			//System.out.println("activationInputs: "+activationInputs[i]);
 		}
 	}
 
@@ -298,20 +376,37 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * its successor links, predecessor links and conflicter links is computed.
 	 */
 	public void computeLinks (){
+        int highest = highestNumPreconditions();
 		for(int i = 0; i < modules.size(); i++){
-			boolean isExecutable = executable(i);
-			activationSuccesors[i] = spreadsForward(modules.get(i).getId(), isExecutable);
-			activationPredeccesors[i] = spreadsBackward(modules.get(i).getId(), isExecutable);
-			activationConflicters[i] = takesAway(modules.get(i).getId());
+			//boolean isExecutable = executable(i);
+            boolean isExecutable = executable(i, highest);
+			activationSuccesors[i] = spreadsForward(modules.get(i).getIdx(), isExecutable);
+			activationPredeccesors[i] = spreadsBackward(modules.get(i).getIdx(), isExecutable);
+			activationConflicters[i] = takesAway(modules.get(i).getIdx());
 		}
 	}
 
-	/**
+    /**
+     * This method returns the behavior with the highest number of preconditions present at the current state.
+     * @return
+     */
+    private int highestNumPreconditions() {
+        int maximum = 0;
+        for( int i = 0; i < modules.size(); i++ ){
+            int numMatched = modules.get( i ).calculateMatchPreconditions( states );
+            if( numMatched > maximum ){
+                maximum = numMatched;
+            }
+        }
+        return maximum;
+    }
+
+    /**
 	 * An executable competence module x spreads activation forward. It increases
 	 * (by a fraction of its own activation level) the activation level of those
 	 * successors y for which the shared proposition p E ax n cy is not true.
 	 * Intuitively, we want these successor modules to become more activated because
-	 * they are `almost executable', since more of their preconditions will be fulfilled
+	 * they are 'almost executable', since more of their preconditions will be fulfilled
 	 * after the competence module has become active.
 	 * Formally, given that competence module x = (cx ax dx zx) is executable,
 	 * it spreads forward through those successor links for which the proposition
@@ -320,18 +415,19 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	public double[] spreadsForward(int indexBehavior, boolean executable){
 		double[] activation = new double[modules.size()];
 		if(executable){
-			BehaviorInterface beh = modules.get(indexBehavior);
-			List<String> addList = new Vector<> (beh.getAddList());
-
-			for(String addPremise : addList){
-				if( !states.contains(addPremise)){ //p E ax is false
-					List<BehaviorInterface> sparseBehaviors = matchProposition(addPremise); //j E ax n cy
-					for(int j = 0; j < sparseBehaviors.size(); j++){
+			Vector<String> addList = new Vector<> ( modules.get(indexBehavior).getAddList() );
+			for(String addPremise : addList ){
+				if( !states.contains(addPremise) ){ //p E ax is false
+                    List<BehaviorPlus> sparseBehaviors = matchProposition(addPremise); //j E ax n cy
+					for(BehaviorPlus beh : sparseBehaviors ){
 						int cardinalityMj = sparseBehaviors.size();
-						int cardinalityCy = sparseBehaviors.get(j).getPreconditions().size();
+						int cardinalityCy = beh.getPreconditions().size();
 						double temp = beh.getActivation() * (phi/gamma) * (1d / cardinalityMj) * (1d / cardinalityCy);
-						activation[sparseBehaviors.get(j).getId()] += temp;
-						//System.out.println(beh.getName()+" spreads "+temp+" forward to "+sparseBehaviors.get(j).getName()+" for "+addList.get(i));
+                        activation[beh.getIdx()] += temp;
+                        if( verbose ) {
+                            System.out.println(modules.get(indexBehavior).getName() + " spreads " + temp + " forward to " +
+                                    beh.getName() + " for " + addPremise);
+                        }
 					}
 				}
 			}
@@ -355,20 +451,26 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	public double[] spreadsBackward(int indexBehavior, boolean executable){
 		double[] activation = new double[modules.size()];
 		if(!executable){
-			BehaviorInterface beh = modules.get(indexBehavior);
-			List<String> condList = new Vector<> (beh.getPreconditions());
+			BehaviorPlus beh = modules.get(indexBehavior);
+			Collection<List<String>> condList = new Vector<> (beh.getPreconditions());
 
-			for(String cond : condList){
-				if(!states.contains(cond)){ //p E cx is false
-					List<BehaviorInterface> sparseBehaviors = achieveProposition(cond); //j E cx n ay
-					for(int j = 0; j < sparseBehaviors.size(); j++){
-						int cardinalityAj = sparseBehaviors.size();
-						int cardinalityAy = sparseBehaviors.get(j).getAddList().size();
-						double temp = beh.getActivation() * (1d / cardinalityAj) * (1d / cardinalityAy);
-						activation[sparseBehaviors.get(j).getId()] += temp;
-						//System.out.println(beh.getName()+" spreads "+temp+" backward to "+sparseBehaviors.get(j).getName()+" for "+condList.get(i));
-					}
-				}
+			for(List<String> condRow : condList ){
+                for(String cond : condRow ){
+                    if( !states.contains( cond ) ) { //p E cx is false
+                        Vector<BehaviorPlus> sparseBehaviors = achieveProposition(cond); //j E cx n ay
+                        for (int j = 0; j < sparseBehaviors.size(); j++) {
+							int cardinalityAj = sparseBehaviors.size();
+                            int cardinalityAy = sparseBehaviors.get(j).getAddList().size();
+                            double temp = beh.getActivation() * (1d / cardinalityAj) * (1d / cardinalityAy);
+                            activation[sparseBehaviors.get(j).getIdx()] += temp;
+                            if( verbose ) {
+                                System.out.println(beh.getName() + " spreads " + temp + " backward to " +
+                                        sparseBehaviors.get(j).getName() + " for " + cond);
+                            }
+                        }
+
+                    }
+                }
 			}
 		}
 		return activation;
@@ -390,24 +492,28 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 */
 	public double[] takesAway(int indexBehavior){
 		double[] activation = new double[modules.size()];
-		BehaviorInterface behx = modules.get(indexBehavior);
-		List<String> condListX = new Vector<>(behx.getPreconditions());
+		BehaviorPlus behx = modules.get(indexBehavior);
+        Collection<List<String>> condListX = new Vector<>(behx.getPreconditions());
 
-		for(String cond : condListX){
-			List<BehaviorInterface> sparseBehaviors = undoPropositionState(cond, indexBehavior); //j E cx n dy
-			for(int j = 0; j < sparseBehaviors.size(); j++){
-				BehaviorInterface behy = sparseBehaviors.get(j);
-				double temp;
-				if((behx.getActivation() <= behy.getActivation())  && inverseTakesAway(indexBehavior, behy.getId())){
-					activation[behy.getId()] = 0;
-				}else{
-					int cardinalityUj = sparseBehaviors.size();
-					int cardinalityDy = sparseBehaviors.get(j).getDeleteList().size();
-					temp = behx.getActivation() * (delta/gamma) * (1d / cardinalityUj) * (1d / cardinalityDy);
-					activation[sparseBehaviors.get(j).getId()] += temp;//Math.max(activationTemp, sparseBehaviors.get(j).getActivationPrior());
-				}
-				//System.out.println(behx.getName()+" decreases (inhibits)" +behy.getName()+" with "+temp+" for "+condListX.get(i));
-			}
+		for( List<String> condRow : condListX ){
+            for( String cond : condRow ) {
+				Vector<BehaviorPlus> sparseBehaviors = undoPropositionState(cond, indexBehavior); //j E cx n dy
+                for ( BehaviorPlus behy : sparseBehaviors) {
+                    double temp = 0;
+                    if ((behx.getActivation() <= behy.getActivation()) && inverseTakesAway(indexBehavior, behy.getIdx())) {
+                        activation[behy.getIdx()] = 0;
+                    } else {
+                        int cardinalityUj = sparseBehaviors.size();
+                        int cardinalityDy = behy.getDeleteList().size();
+                        temp = behx.getActivation() * (delta / gamma) * (1d / cardinalityUj) * (1d / cardinalityDy);
+                        activation[behy.getIdx()] += temp;//Math.max(activationTemp, sparseBehaviors.get(j).getActivationPrior());
+                    }
+                    if( verbose ) {
+                        System.out.println(behx.getName() + " decreases (inhibits)" + behy.getName() + " with " + temp +
+                                " for " + cond);
+                    }
+                }
+            }
 		}
 		return activation;
 	}
@@ -419,13 +525,15 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * @return
 	 */
 	private boolean inverseTakesAway(int indexBehx, int indexBehy){
-		List<String> condsListy = new Vector<> (modules.get(indexBehy).getPreconditions());
-		BehaviorInterface behx = modules.get(indexBehx);
+        Collection<List<String>> condsListy = new Vector<> (modules.get(indexBehy).getPreconditions());
+		BehaviorPlus behx = modules.get(indexBehx);
 
-		for(String cond : condsListy){
-			if( behx.getDeleteList().contains(cond) ) //cy n dx
-				if( states.contains(cond) ) //cy n S(t)
-					return true;
+		for(List<String> condRow : condsListy ){
+            for( String cond : condRow ) {
+                if (behx.getDeleteList().contains(cond) ) //cy n dx
+                    if (states.contains( cond )) //cy n S(t)
+                        return true;
+            }
 		}
 		return false;
 	}
@@ -446,17 +554,17 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 		}
 		if(sum > pi * modules.size()){
 			factor = pi * modules.size() / sum;
-			for(BehaviorInterface beh : modules){
-				beh.decay(factor);
+			for(int i = 0; i < modules.size(); i++){
+				modules.get(i).decay(factor);
 			}
 		}
 		//ojrl decay bahavior which increments its activation continously and is not activated
 		if(modules.get(indexMayor).getActivation() > (theta * 1.5) && !modules.get(indexMayor).getActivated()){
 			modules.get(indexMayor).decay(0.5);
-			for (BehaviorInterface module : modules) {
-				if (module.getExecutable() && module.getId() != modules.get(indexMayor).getId()
+			for (BehaviorPlus module : modules) {
+				if (module.getExecutable() && module.getIdx() != modules.get(indexMayor).getIdx()
 						&& !module.getActivated() && indexBehActivated != -1) {
-					if (indexBehActivated != module.getId())
+					if (indexBehActivated != module.getIdx())
 						module.decay(1.5);
 					else
 						module.decay(0.9);
@@ -465,7 +573,9 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 		}
 
 		for(int i = 0; i < modules.size(); i++){
-			//System.out.println("activation-level "+modules.get(i).getName()+": "+modules.get(i).getActivation());
+            if( verbose ) {
+//                System.out.println("activation-level " + modules.get(i).getName() + ": " + modules.get(i).getActivation());
+            }
 		}
 	}
 
@@ -480,63 +590,101 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * threshold is lowered by 10%.
 	 */
 	public int activateBehavior (){
-		boolean[] active = new boolean[modules.size()];
 		double act = 0;
-		int index = 0;
 		indexBehActivated = -1;
-		for(int j = 0; j < modules.size(); j++){
-			if(modules.get(j).getActivation() > act){
-				index = j;
-				act = modules.get(j).getActivation();
+		for( BehaviorPlus beh : modules ){
+            if( beh.getActivation() >= theta && beh.getExecutable() && beh.getActivation() > act){
+                indexBehActivated = modules.indexOf(beh);
+                act = beh.getActivation();
+                nameBehaviorActivated = beh.getId(); //beh.getName();
+            }
+        }
+        execution = false;
+        if( indexBehActivated >= 0 ){
+			statesOutput = Arrays.toString(states.toArray());
+			if(SocialReasonerController.verbose ) {
+				System.err.println("Executing Behavior: " + modules.get(indexBehActivated).getName());
+				System.err.println("States: " + statesOutput );
 			}
-		}
-
-		for(int i = 0; i < modules.size(); i++){
-			BehaviorInterface beh = modules.get(i);
-			if(beh.getActivation() >= theta && beh.getExecutable()){
-				active[i] = i == index;
-			}else
-				active[i] = false;
-		}
-		execution = false;
-		for(int i = 0; i < modules.size(); i++){
-			if( active[i] ){
-				System.err.println("Executing Behavior "+modules.get(i).getName());
-				execution = true;
-				indexBehActivated = i; //modules.get(i).getIdx();
-				modules.get(i).setActivated(true);
-				protectGoals(modules.get(i));
-				execute(indexBehActivated);
-				break; //ojrl: determine if there are several behaviors activated in the same step, then apply stochastic method
+			String values = "";
+			matchesOutput = "";
+			for( BehaviorPlus beh : modules ){
+				String matches = "$ " + beh.getName() + ": ";
+				int number = 0;
+				for( String state : states){
+					for( List<String> precond : beh.getPreconditions() ){
+						if( precond.contains(state) ){
+							matches += "[" + state + "] ";
+							number++;
+						}
+					}
+				}
+				if( SocialReasonerController.verbose ) {
+					System.err.println(matches.replace("$ ", "(" + number + "-" + beh.getActivation() + ") "));
+				}
+				matchesOutput += matches.replace("$ ", "(" + number + ") ") + ", ";
+				values += "["+beh.getName() + ": " + beh.getActivation() + "], ";
 			}
-		}
-		if( !execution ){
-			System.err.println("None of the executable modules has accumulated enough activation to become active");
-			theta *= 0.9;
-			System.out.println("Theta: "+theta);
-		}
+			if( SocialReasonerController.verbose ) {
+				System.out.println(values + ", theta: " + theta);
+			}
+            execution = true;
+            modules.get(indexBehActivated).setActivated(true);
+            protectGoals(modules.get(indexBehActivated));
+        }
 		return indexBehActivated;
 	}
 
 	/**
-	 * Execute the behavior
+	 *
 	 * @param behIndex
 	 */
-	public void execute(int behIndex){
-		BehaviorInterface beh = modules.get(behIndex);
-		List<String> addList = new Vector<>(beh.getAddList());
-		List<String> deleteList = new Vector<>(beh.getDeleteList());
+	public synchronized void triggerPostconditions(int behIndex){
+		if( behIndex >= 0 && behIndex < modules.size() ) {
+			BehaviorPlus beh = modules.get(behIndex);
+			Vector<String> addList = new Vector<>(beh.getAddList());
+			Vector<String> deleteList = new Vector<>(beh.getDeleteList());
 
-		for (String anAddList : addList) {
-			states.remove( anAddList );
-			states.add(anAddList);
-		}
-		for (String aDeleteList : deleteList) {
-			states.remove(aDeleteList);
-		}
-		for (String goal : beh.getAddGoals()) {
-			goals.remove( goal );
-			goals.add( goal );
+			synchronized (states) {
+				try {
+					previousStates = new Vector<>(states);
+					//we do not need to add preconditions to state since it is done by DMMain
+					//                for (String anAddList : addList) {
+					//                    if (!anAddList.contains("_history")) {
+					//                        states.remove(anAddList);
+					//                        states.add(anAddList);
+					//                    }
+					//                }
+					if (removePrecond) {
+						if( beh.getPreconditions() != null ) {
+							for (List<String> preconds : beh.getPreconditions()) {
+								for (String precond : preconds) {
+									states.remove(precond);
+								}
+							}
+						}
+					} else {
+						List<String> toRemove = new ArrayList<>();
+						for (String premise : deleteList) {
+							for (String st : states) {
+								if (premise.contains("*") && Pattern.compile(premise.replace("*", "[a-zA-Z0-9]*"))
+										.matcher(st).matches()) {
+									toRemove.add(st);
+								}
+							}
+							for (String st : toRemove) {
+								states.remove(st);
+							}
+						}
+					}
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+			for (String goal : beh.getAddGoals()) {
+				goals.remove(goal);
+				goals.add(goal);
+			}
 		}
 	}
 
@@ -544,7 +692,7 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 	 * Protect the goals achieved
 	 * @param beh
 	 */
-	public void protectGoals(BehaviorInterface beh){
+	public void protectGoals(BehaviorPlus beh){
 		for (String goalTemp : beh.getAddList()) {
 			if (goals.contains(goalTemp)) {
 				goalsResolved.add(goalTemp);
@@ -575,49 +723,146 @@ public class BehaviorNetwork implements BehaviorNetworkInterface{
 		//reseting
 		if(execution)
 			theta = initialTheta;
-		for (BehaviorInterface module : modules) {
+		for (BehaviorPlus module : modules) {
 			module.resetActivation(module.getActivated());
 		}
 	}
 
-	public int selectBehavior() {
-		activations[ modules.size() ] = theta;
-		System.out.println("\nStep "+(cont+1)+" goals achieved: " + goalsResolved.size() + " goals: " + goals.size());
-		List<String> st = new Vector<>( states);
-		String state = "";
-		for (String aSt : st) {
-			state += aSt + "  ";
+    public int selectBehavior() {
+        if (activations == null) {
+            setModules(modules, modules.size() + 1);
+        }
+        activations[modules.size()] = theta;
+        if (verbose) {
+            System.out.println("\nStep " + (cont + 1) + " goals achieved: " + goalsResolved.size() + " goals: " + goals.size());
+        }
+        computeActivation();
+
+        computeLinks();
+
+        updateActivation();
+
+        activateBehavior();
+
+        recordActivations();
+
+        applyDecayFx();
+
+        cont++;
+        return indexBehActivated;
+    }
+
+    private void recordActivations(){
+        for (int i = 0; i < modules.size(); i++) {
+            BehaviorPlus b = modules.get(i);
+            if (verbose) {
+                System.out.println("Behavior: " + b.getName() + "  activation: " + b.getActivation());
+            }
+            activations[i] = modules.get(i).getActivation();
+        }
+        if (indexBehActivated != -1) {
+            activations[modules.size()] = modules.get(indexBehActivated).getActivation();
+        }
+    }
+
+    public void execute(int cycle){
+        if( cycle > 0) {
+            triggerPostconditions(indexBehActivated);
+        }
+        if( !execution ){
+            theta *= 0.9;
+            if( verbose ) {
+                System.err.println("None of the executable modules has accumulated enough activation to become active");
+                System.out.println("Theta: " + theta);
+            }
+        }
+
+        if( cycle > 0 ) {
+            reset();
+        }
+    }
+
+    public int getHighestActivation() {
+        int idx = 5; //ASN
+        double max = 0;
+        for(int i = 0; i< modules.size(); i++){
+            BehaviorPlus beh = modules.get(i);
+            if( beh.getActivation() > theta && beh.getActivation() > max ){
+                idx = i;
+            }
+        }
+        modules.get(idx).setActivation(max * 1.15);
+        nameBehaviorActivated = modules.get(idx).getId(); //.getName();
+        return indexBehActivated = idx;
+    }
+
+	public int getHighestActivationUsingNone(){
+		int idx = 7; //NONE
+		double max = 0;
+		for(int i = 0; i< modules.size(); i++){
+			BehaviorPlus beh = modules.get(i);
+			if( beh.getName().equals("NONE")){
+				idx = i;
+			}
+			if( beh.getActivation() > theta && beh.getActivation() > max ){
+				max = beh.getActivation();
+			}
 		}
-		System.out.println("State: "+state);
+        if( max == 0){
+            max = theta;
+        }
+        indexBehActivated = idx;
+        nameBehaviorActivated = modules.get(idx).getId(); //getName();
+        modules.get(idx).setActivation(max * 1.30);
+        modules.get(idx).setActivated(true);
+        execution = true;
+        recordActivations();
+		return idx;
+	}
 
-		// 1
-		computeActivation();
+    public String[] getModuleNamesByHighestActivation() {
+        getModulesCopy();
+        Collections.sort( modulesCopy );
+        Collections.reverse( modulesCopy );
+        String[] results = new String[modulesCopy.size()];
+		DecimalFormat formatter = new DecimalFormat("#0.00");
+		output = "";
+        for( int i = 0; i < modulesCopy.size(); i++ ){
+            results[i] = modulesCopy.get(i).getId();
+			output += "["+modulesCopy.get(i).getName()+": "+formatter.format( modulesCopy.get(i).getActivation())+"], ";
+        }
+		output += "theta: " + theta;
+        return results;
+    }
 
-		// 2
-		computeLinks();
+    public void getModulesCopy() {
+        modulesCopy.clear();
+        for( BehaviorPlus beh : modules ){
+            modulesCopy.add( beh.clone() );
+        }
+    }
 
-		// 4
-		updateActivation();
-
-		// 3
-		activateBehavior();
-
-		System.out.println("\n");
-		for( int i = 0; i < modules.size(); i++ ){
-			BehaviorInterface b = modules.get(i);
-			System.out.println("Behavior: " + b.getName() + "  activation: "+b.getActivation());
-			activations[i] = modules.get(i).getActivation();
+	public String[] getModuleNames() {
+		String[] names = new String[ modules.size() ];
+		for(int i = 0; i < names.length; i++){
+			names[i] = modules.get(i).getName();
 		}
-		if( indexBehActivated != -1 ) {
-			activations[modules.size()] = modules.get(indexBehActivated).getActivation();
+		return names;
+	}
+
+	public double[] getOnlyActivations() {
+		int size = modules.size();
+		double[] actvs = new double[ size ];
+		for( int i = 0; i < size; i++){
+			actvs[i] = activations[i];
 		}
+		return actvs;
+	}
 
-		// 5
-		applyDecayFx();
-
-		//6
-		reset();
-		cont++;
-		return indexBehActivated;
+	public void resetAll() {
+		for( BehaviorPlus beh : modules ){
+			beh.reset();
+		}
+		theta = initialTheta;
 	}
 }
