--- src/main/java/edu/cmu/inmind/multiuser/socialreasoner/control/bn/InterfaceAdapter.java
@@ -1,40 +1,0 @@
-package edu.cmu.inmind.multiuser.socialreasoner.control.bn;
-
-import com.google.gson.*;
-
-import java.lang.reflect.Type;
-
-/**
- * Created by oscarr on 5/25/16.
- */
-public final class InterfaceAdapter<T> implements JsonSerializer<T>, JsonDeserializer<T> {
-    public JsonElement serialize(T object, Type interfaceType, JsonSerializationContext context) {
-        final JsonObject wrapper = new JsonObject();
-        wrapper.addProperty("BehaviorNetworkPlus", object.getClass().getName());
-        wrapper.add("data", context.serialize(object));
-        return wrapper;
-    }
-
-    public T deserialize(JsonElement elem, Type interfaceType, JsonDeserializationContext context) throws JsonParseException {
-        final JsonObject wrapper = (JsonObject) elem;
-        final JsonElement typeName = get(wrapper, "modules");
-        final JsonElement data = get(wrapper, "List<BehaviorInterface>");
-        final Type actualType = typeForName(typeName);
-        return context.deserialize(data, actualType);
-    }
-
-    private Type typeForName(final JsonElement typeElem) {
-        try {
-            return Class.forName(typeElem.getAsString());
-        } catch (ClassNotFoundException e) {
-            throw new JsonParseException(e);
-        }
-    }
-
-    private JsonElement get(final JsonObject wrapper, String memberName) {
-        final JsonElement elem = wrapper.get(memberName);
-        if (elem == null) throw new JsonParseException("no '" + memberName
-                + "' member found in what was expected to be an interface wrapper");
-        return elem;
-    }
-}
