--- src/main/java/edu/cmu/inmind/multiuser/socialreasoner/control/bn/MainTestBNPlus.java
@@ -1,42 +1,0 @@
-package edu.cmu.inmind.multiuser.socialreasoner.control.bn;
-
-/**
- * Created by oscarr on 5/12/16.
- */
-public class MainTestBNPlus {
-    public static void main(String args[]){
-//        Vector<BehaviorInterface> modules = new Vector();
-//
-//        modules.add( new BehaviorPlus( "A", "A",
-//                new String[][]{new String[]{"P1", "P1.1", "P1.2", "P1.3"}, new String[]{"P2", "P2.1", "P2.2", "P2.3"}},
-//                new String[]{"P3", "P4.2"},
-//                new String[]{"P1", "P1.1", "P1.2", "P1.3", "P2", "P2.1", "P2.2", "P2.3"}) );
-//
-//        modules.add( new BehaviorPlus( "B", "B",
-//                new String[][]{new String[]{"P3", "P3.1", "P3.2", "P3.3"}, new String[]{"P4", "P4.1", "P4.2", "P4.3"}},
-//                new String[]{"P5.3", "P6"},
-//                new String[]{"P3", "P3.1", "P3.2", "P3.3", "P4", "P4.1", "P4.2", "P4.3"}) );
-//
-//        modules.add( new BehaviorPlus( "C", "C",
-//                new String[][]{new String[]{"P5", "P5.1", "P5.2", "P5.3"}, new String[]{"P6", "P6.1", "P6.2", "P6.3"}},
-//                new String[]{"P7"},
-//                new String[]{"P5", "P5.1", "P5.2", "P5.3", "P6", "P6.1", "P6.2", "P6.3"}) );
-//
-//        BehaviorNetworkPlus bnp = new BehaviorNetworkPlus();
-//        bnp.setModules(modules);
-//        bnp.setState( Arrays.asList( new String[]{"P1.2", "P2.3"}));
-//        bnp.setGoals( Arrays.asList( new String[]{"P7"} ) );
-//        bnp.setPi(20);
-//        bnp.setTheta(15);
-//        bnp.setInitialTheta(15);
-//        bnp.setPhi(70);
-//        bnp.setGamma(20);
-//        bnp.setDelta(50);
-//        bnp.setModules(modules, modules.size() + 1);
-//        bnp.setGoalsR(new Vector<>());
-//
-//        while( bnp.getGoalsR().size() < bnp.getGoals().size() ){
-//            bnp.selectBehavior();
-//        }
-    }
-}
