+++ stderr.2017-08-14T16:02:47
@@ -25,8 +25,8 @@ java.lang.ArrayIndexOutOfBoundsException: -1
 	at edu.cmu.inmind.multiuser.socialreasoner.control.bn.BehaviorNetworkPlus.triggerPostconditions(BehaviorNetworkPlus.java:636)
 	at edu.cmu.inmind.multiuser.socialreasoner.control.bn.BehaviorNetworkPlus.execute(BehaviorNetworkPlus.java:759)
 	at edu.cmu.inmind.multiuser.socialreasoner.control.reasoners.SocialReasoner.execute(SocialReasoner.java:106)
-	at edu.cmu.inmind.multiuser.socialreasoner.control.MainController.process(MainController.java:116)
-	at edu.cmu.inmind.multiuser.socialreasoner.control.MainController.addSystemIntent(MainController.java:405)
+	at edu.cmu.inmind.multiuser.socialreasoner.control.SocialReasonerController.process(SocialReasonerController.java:116)
+	at edu.cmu.inmind.multiuser.socialreasoner.control.SocialReasonerController.addSystemIntent(SocialReasonerController.java:405)
 	at edu.cmu.inmind.multiuser.sara.component.SocialReasonerComponent.selectStrategy(SocialReasonerComponent.java:97)
 	at edu.cmu.inmind.multiuser.sara.component.SocialReasonerComponent.onEvent(SocialReasonerComponent.java:137)
 	at edu.cmu.inmind.multiuser.sara.component.SocialReasonerComponent$$EnhancerByGuice$$d4bb2d5.CGLIB$onEvent$0(<generated>)
